refactor(solver): further-split index_access.rs to 1794 lines + ratchet-down
Behavior-preserving extraction of the optional-property / mapped optional-read / constraint-resolved-index helper cluster into a new sibling module index_access_optional_read.rs. index_access.rs drops from 2033 to 1794 lines (262 under the 2000 hard ceiling) and the arch ratchet is lowered to match. No semantic change.
